+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Features:
 * [#268](https://github.com/godaddy/tartufo/issues/268) - Adds a new
   `--recurse / --no-recurse` flag which allows users to recursively scan the entire directory or just
   the root directory
+* [#202](https://github.com/godaddy/tartufo/issues/202) - Supports new format of exclusions in config file 
+  with the ability to specify the reason along with exclusion
 
 v3.0.0-alpha.1 - 11 November 2021
 ---------------------------------